Print characters 'c with high bit set as 'c|$80; bump version
This is very useful for programs that set the high bit of characters
to indicate end of string or for some other reason. Sample output:

    FCB     'E,'N,'D|$80             ;015C: 45 4E C4       'EN.'

This applies only to memory locations explicitly marked as `char`
(in the info file). If you're doing a disassembly later to be
assembled by an assembler that doesn't support this "logical or"
syntax you can either not mark those bytes as `char` or, if you've
marked a large range like `char 015C-0233`, add additional directives
after that to mark individual bytes as non-char, e.g., `hex 015E` to
would produce `$C4` instead of `'D|$80` in the output above.

This syntax is supported by AS and A09, but does not appear to be
supported by TSC.